use simple Middleware call
